+++ src/lib.rs
@@ -50,6 +50,23 @@
 //! ## Integration
 //!
 //! You can find an example of integration in the [common module](examples/common/mod.rs) of the examples.
+//! 
+//! ```rust
+//! // Example with default allocator
+//! let renderer = Renderer::with_default_allocator(
+//!     &instance,
+//!     physical_device,
+//!     device.clone(),
+//!     graphics_queue,
+//!     command_pool,
+//!     render_pass,
+//!     &mut imgui,
+//!     Some(Options {
+//!         in_flight_frames: 1,
+//!         ..Default::default()
+//!     }),
+//! ).unwrap();
+//! ```
 //!
 //! ## Examples
 //!