+++ src/renderer/allocator/gpu.rs
@@ -142,9 +142,9 @@ impl Allocate for Allocator {
         unsafe {
             let data_ptr = memory
                 .mapped_ptr()
-                .ok_or(RendererError::Allocator(
-                    "Failed to get mapped memory pointer".into(),
-                ))?
+                .ok_or_else(|| {
+                    RendererError::Allocator("Failed to get mapped memory pointer".into())
+                })?
                 .as_ptr();
             let mut align = ash::util::Align::new(data_ptr, std::mem::align_of::<T>() as _, size);
             align.copy_from_slice(data);